System.IO.Posix.MMap.Internal

Description

Low level mmap access.

Converting an mmapped pointer to a ByteString

unsafePackMMapPtr :: Ptr Word8 -> CSize -> IO ByteString

Create a bytestring from a memory mapped Ptr. A finalizer will be associated with the resource, that will call munmap when the storage manager detects that the resource is no longer in use.

Low level bindings

c_mmap :: CSize -> CInt -> IO (Ptr Word8)

c_munmap :: Ptr Word8 -> CSize -> IO CInt